+++ src/ui_egui/app/views/mod.rs
@@ -625,7 +625,7 @@ impl CalendarApp {
 
 /// Shift a date by the given number of months
 fn shift_month(date: NaiveDate, delta: i32) -> NaiveDate {
-    let total_months = (date.year() * 12) as i32 + (date.month() as i32 - 1) + delta;
+    let total_months = (date.year() * 12) + (date.month() as i32 - 1) + delta;
     let new_year = total_months.div_euclid(12);
     let new_month = (total_months.rem_euclid(12) + 1) as u32;
     let max_day = days_in_month(new_year, new_month);
